Story time: Going down the Jekyll-Lunr Rabbit Hole
Background: Jekyll-Lunr provides search for pure static content without
a backend.

Kenny pointed out that the new webserver-embedded devcenter
had an error in the JS console: `TypeError: $(...).modal is not a
function`. The specific call was for popping up our TOS if the modal was
rendered, but completely irrelevant to the devcenter. But the error
didn't happen on pages outside the devcenter, so it was indicative of a
collision.

After a bit of debugging, I discovered that the Jekyll-Lunr
plugin's `search.min.js` was crushing the webserver's JQuery, because
the plugin was bundling JQuery, so we were effectively loading
all the JS for the webserver, and then reloading JQuery from scratch
before the page was ready.

I attempted to prune jquery from `search.min.js`, but that was a
spectacular failure. Rather than add jquery.lunr.search.js and 5 out of
its 6 dependencies individually and add an asset pipeline to Jekyll, I
just cloned the the project, commented out `jquery` from the build step
and created a jquery-free `search.min.js`.

And another JS error appeared. It turns out the plugin we were using
used an old lunr.js index format, so my new `search.min.js` was
incompatible, so naturally rather than go build an outdated
`search.min.js`, I opted for upgrading the jekyll plugin.

Viola! No JS errors, but now the search results didn't load correctly.
A minor change in the API was easy to fix, but then I realized that I
had previously hacked the jekyll-lunr gem to also print the `excerpt` of
a search result, after all it's nice to see a little snippet about each search
result (I think I'll submit it as a PR since I think it's a nice feature). So,
a bit more tweaking to port that change into the updated gem and success!