minor typos and suggestions in python client docs
note that the following changes are per PEP-8 style guidelines:

`if not img_directory.exists():` instead of `if img_directory.exists() is False:`
`if not client.file(img_file).exists()` instead of `if client.file(img_file).exists() is False`
`if client.file(download_uri).exists():` instead of `if client.file(download_uri).exists() is True:`

Also to note: in its current state, the algoCV algorithm ('dlib/FaceDetection') does not accept a `pixels` value in the input; maybe this was part of a previous version of the algorithm, in which case that version needs to be pinned by adding the version number to the `client.algo()` call. Also, I do not currently get an 'output' key in the response object, only a `url` key; maybe that was also part of the previous algorithm version.